// backend/src/types/api.d.ts
export interface components {
    schemas: {
// ...
        ConversionEventCreateRequest: {
            call_public_id: string;
            buyer_id: string;
            /** @example CASE_FILED */
            event_type: string;
            /** Format: date-time */
            event_time: string;
            revenue_cents?: number | null;
        };
        ConversionEventCreateResponse: {
            id: string;
            call_session_id: string;
            /** @enum {string} */
            status: "ACCEPTED";
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}